Move Projects and Environments into their own files
