+++ atst/domain/environments.py
@@ -1,0 +1,12 @@
+from atst.database import db
+from atst.models.environment import Environment
+
+
+class Environments(object):
+    @classmethod
+    def create(cls, project, name):
+        environment = Environment(project=project, name=name)
+        db.session.add(environment)
+        db.session.commit()
+        return environment
+
